requirements.txt
Flask_SocketIO==4.0.0
Flask==1.0.2
pymongo==3.4.0
python-dateutil==2.8.1
pytz==2020.1
gunicorn==19.9.0
eventlet==0.28.0